refactor: consolidate redirection-hosts shared types
Extract EmailFieldProps and RedirectionSslValues into redirectionFormTypes.ts
to eliminate duplication between LetsEncryptForm and RedirectionSslSection.

Part of #103
