+++ src/components/features/redirection-hosts/RedirectionSslSection.tsx
@@ -4,37 +4,12 @@ import {
   Switch,
   Typography,
 } from '@mui/material'
-import type { ChangeEvent } from 'react'
 import { Certificate } from '../../../api/certificates'
 import CertificateSelector from '../../shared/CertificateSelector'
 import LetsEncryptForm from './LetsEncryptForm'
+import type { EmailFieldProps, RedirectionSslValues } from './redirectionFormTypes'
 
-export interface RedirectionSslValues {
-  certificate_id: number | string
-  use_lets_encrypt: boolean
-  ssl_forced: boolean
-  hsts_enabled: boolean
-  hsts_subdomains: boolean
-  http2_support: boolean
-  letsencrypt_email: string
-  letsencrypt_agree: boolean
-  dns_challenge: boolean
-  dns_provider: string
-  dns_provider_credentials: string
-  propagation_seconds: string
-}
-
-/** Full shape returned by useDrawerForm.getFieldProps — passed through to preserve
- *  onBlur (touch tracking), disabled (while submitting), and touch-gated error display. */
-export interface EmailFieldProps {
-  name: string
-  value: string | number | readonly string[]
-  onChange: (event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => void
-  onBlur: () => void
-  error: boolean
-  helperText?: string
-  disabled: boolean
-}
+export type { RedirectionSslValues }
 
 interface RedirectionSslSectionProps {
   values: RedirectionSslValues
